requirements.txt
asgiref==3.6.0
Django==4.1.4
django-cleanup==6.0.0
django-shopping-cart==0.1
Pillow==9.3.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
